Finally got day 23 working - both parts!
Solved as a graph of graphs!! So the individual movement of amphipods is solved using a-star, then each possible move is held as a node in a graph and DFS used to calculate the shortest path... eep!

Also did work in class comparing relative merits of dijkstra implementations (then compared against a-star) on day 15.