Bit of a catch-up - days 7, 8, and 9 complete
Nothing hugely concerning, here. Pleased with how it's gone - did enjoy the camel poker of day 7 and was pleased with how I'd tackled part 1, so part 2 played straight into it; day 8 part 2 was an "a-ha" moment, having tried brute force - I wouldn't have know how to prove it *was* going to work, simply that it was the next thing to try when brute force plainly wasn't going to. And then today's was fun, although actually easier than I expected - maybe it was harder if you over-thought the maths of it?