Blinking day 12
Finally succumbed to finding out how to sort part 2 and followed an explanation on the Reddit group which talked through how to use DP recursion... have to say that it made a whole heap of sense, in terms of approach. Now have to see if I can managed to apply that elsewhere, going forwards.