Day 17 finally complete
It took a *long* while, but I finally have day 17 working (not quickly, but it works) - part 1 runs in less than a minute using A*... but part 2 can take over 10 minutes (I'm posting to ask for tips on Reddit).

The key was around the neighbours, and understanding how they were to be returned - very intricate and unlike anything I'd seen before. Once I finally got my head around part 1, part 2 wasn't that big a leap... it's just a long run to completion (which is what I'm hoping for tips on).