Day 23 performance improvements
Bit of a tidy up meaning that I halved the timings for the DFS on the part 2, which is nice. Largely by actually storing the path length as part of the actual traversal which means that (a) the visited can be stored as a set again, and (b) the total cost can be returned in one go without having to retrace the path afterwards. All good.